+++ components/atom/actionButton/test/index.test.js
@@ -46,6 +46,19 @@ describe('atom/actionButton', () => {
     expect(container.innerHTML).to.not.have.lengthOf(0)
   })
 
+  it.skip('should NOT extend classNames', () => {
+    // Given
+    const props = {className: 'extended-classNames'}
+    const findSentence = str => string => string.match(new RegExp(`S*${str}S*`))
+
+    // When
+    const {container} = setup(props)
+    const findClassName = findSentence(props.className)
+
+    // Then
+    expect(findClassName(container.innerHTML)).to.be.null
+  })
+
   it('Displays the expected button', () => {
     // Given
     const props = {