+++ components/atom/pinInput/test/index.test.js
@@ -49,6 +49,19 @@ describe('AtomPinInput', () => {
     expect(container.innerHTML).to.not.have.lengthOf(0)
   })
 
+  it('should NOT extend classNames', () => {
+    // Given
+    const props = {className: 'extended-classNames'}
+    const findSentence = str => string => string.match(new RegExp(`S*${str}S*`))
+
+    // When
+    const {container} = setup(props)
+    const findClassName = findSentence(props.className)
+
+    // Then
+    expect(findClassName(container.innerHTML)).to.be.null
+  })
+
   describe('pinInputMask', () => {
     describe('NUMBER', () => {
       it('should return FALSE when value is undefined', () => {