+++ components/atom/skeleton/test/index.test.js
@@ -37,6 +37,19 @@ describe('AtomSkeleton', () => {
     expect(container.innerHTML).to.not.have.lengthOf(0)
   })
 
+  it('should NOT extend classNames', () => {
+    // Given
+    const props = {className: 'extended-classNames'}
+    const findSentence = str => string => string.match(new RegExp(`S*${str}S*`))
+
+    // When
+    const {container} = setup(props)
+    const findClassName = findSentence(props.className)
+
+    // Then
+    expect(findClassName(container.innerHTML)).to.be.null
+  })
+
   it('should render a set of skeletons', () => {
     // Given
     const count = 3