+++ components/atom/tag/test/index.test.js
@@ -47,6 +47,19 @@ describe('atom/tag', () => {
     expect(container.innerHTML).to.not.have.lengthOf(0)
   })
 
+  it('should NOT extend classNames', () => {
+    // Given
+    const props = {className: 'extended-classNames', label: 'label'}
+    const findSentence = str => string => string.match(new RegExp(`S*${str}S*`))
+
+    // When
+    const {container} = setup(props)
+    const findClassName = findSentence(props.className)
+
+    // Then
+    expect(findClassName(container.innerHTML)).to.be.null
+  })
+
   it('should not trigger click when disabled', () => {
     const spy = sinon.spy()
     const {getByRole} = setup({