+++ components/atom/skeleton/CHANGELOG.md
@@ -1,18 +1,49 @@
-# Change Log
+# CHANGELOG
 
-All notable changes to this project will be documented in this file.
+# 1.4.0 (2021-10-06)
+
+
+### Bug Fixes
+
+* **components/atom/skeleton:** classNames closed ([1e4602d](https://github.com/SUI-Components/sui-components/commit/1e4602d29c0bcc5b257a51d8ecd5dced3302645b))
+
+
+### Features
+
+* **Root:** normalize bborder variables ([b5da148](https://github.com/SUI-Components/sui-components/commit/b5da1482ca96b523f0c168c7040783ce78a7f14d))
+* **Root:** normalize border variables ([13c1d9c](https://github.com/SUI-Components/sui-components/commit/13c1d9ca934d68811f8dccfedb7f74ee834f6930))
+
+
+
+# 1.3.0 (2021-01-05)
+
+
+### Bug Fixes
+
+* **atom/skeleton:** proptypes ([3641c3c](https://github.com/SUI-Components/sui-components/commit/3641c3c4d58a7c671a48b20d51e801b7ab8adc71))
+
+
+
+# 1.2.0 (2020-11-24)
+
+
+### Features
+
+* **atom/skeleton:** Use new jsx runtime ([789c52e](https://github.com/SUI-Components/sui-components/commit/789c52e565c891a1be4f2a7d30ad3bafebfcc06f))
+
+
 
 # 1.1.0 (2020-09-16)
 
 
 ### Bug Fixes
 
-* fix some typos ([a1a823f](https://github.com/SUI-Components/sui-components/commit/a1a823f6b03245c91c6eb4ae80fcad197d576c35))
+* **atom/skeleton:** fix some typos ([a1a823f](https://github.com/SUI-Components/sui-components/commit/a1a823f6b03245c91c6eb4ae80fcad197d576c35))
 
 
 ### Features
 
-* add first iteration of skeleton component ([d435a84](https://github.com/SUI-Components/sui-components/commit/d435a847c9e44aae3615a0ea47d8d02d037375c5))
+* **atom/skeleton:** add first iteration of skeleton component ([d435a84](https://github.com/SUI-Components/sui-components/commit/d435a847c9e44aae3615a0ea47d8d02d037375c5))